+++ src/vs/platform/storage/common/storage.ts
@@ -10,7 +10,7 @@ import { mark } from 'vs/base/common/performance';
 import { isUndefinedOrNull } from 'vs/base/common/types';
 import { InMemoryStorageDatabase, IStorage, Storage } from 'vs/base/parts/storage/common/storage';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
-import { IWorkspaceInitializationPayload } from 'vs/platform/workspaces/common/workspaces';
+import { IAnyWorkspaceIdentifier } from 'vs/platform/workspace/common/workspace';
 
 export const IS_NEW_KEY = '__$__isNewStorageMarker';
 const TARGET_KEY = '__$__targetStorageMarker';
@@ -141,7 +141,7 @@ export interface IStorageService {
 	/**
 	 * Migrate the storage contents to another workspace.
 	 */
-	migrate(toWorkspace: IWorkspaceInitializationPayload): Promise<void>;
+	migrate(toWorkspace: IAnyWorkspaceIdentifier): Promise<void>;
 
 	/**
 	 * Whether the storage for the given scope was created during this session or
@@ -510,7 +510,7 @@ export abstract class AbstractStorageService extends Disposable implements IStor
 
 	protected abstract getLogDetails(scope: StorageScope): string | undefined;
 
-	abstract migrate(toWorkspace: IWorkspaceInitializationPayload): Promise<void>;
+	abstract migrate(toWorkspace: IAnyWorkspaceIdentifier): Promise<void>;
 }
 
 export class InMemoryStorageService extends AbstractStorageService {
@@ -535,7 +535,7 @@ export class InMemoryStorageService extends AbstractStorageService {
 
 	protected async doInitialize(): Promise<void> { }
 
-	async migrate(toWorkspace: IWorkspaceInitializationPayload): Promise<void> {
+	async migrate(toWorkspace: IAnyWorkspaceIdentifier): Promise<void> {
 		// not supported
 	}
 }
