+++ src/vs/platform/storage/electron-main/storageMainService.ts
@@ -13,8 +13,7 @@ import { ILifecycleMainService, LifecycleMainPhase } from 'vs/platform/lifecycle
 import { ILogService } from 'vs/platform/log/common/log';
 import { AbstractStorageService, IStorageService, StorageScope, StorageTarget } from 'vs/platform/storage/common/storage';
 import { GlobalStorageMain, IStorageMain, IStorageMainOptions, WorkspaceStorageMain } from 'vs/platform/storage/electron-main/storageMain';
-import { IWorkspaceInitializationPayload } from 'vs/platform/workspaces/common/workspaces';
-import { IEmptyWorkspaceIdentifier, ISingleFolderWorkspaceIdentifier, IWorkspaceIdentifier } from 'vs/platform/workspace/common/workspace';
+import { IAnyWorkspaceIdentifier, IEmptyWorkspaceIdentifier, ISingleFolderWorkspaceIdentifier, IWorkspaceIdentifier } from 'vs/platform/workspace/common/workspace';
 
 //#region Storage Main Service (intent: make global and workspace storage accessible to windows from main process)
 
@@ -185,7 +184,7 @@ export interface IGlobalStorageMainService extends IStorageService {
 
 	keys(scope: StorageScope.GLOBAL, target: StorageTarget): string[];
 
-	migrate(toWorkspace: IWorkspaceInitializationPayload): never;
+	migrate(toWorkspace: IAnyWorkspaceIdentifier): never;
 
 	isNew(scope: StorageScope.GLOBAL): boolean;
 }
